--- src/matter/interaction/InteractionErrors.ts
@@ -1,20 +1,0 @@
-/**
- * @license
- * Copyright 2022 The node-matter Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import {MatterError} from "../../error/MatterError";
-import {StatusCode} from "./InteractionMessages";
-
-/** Error base Class for all errors related to the status response messages. */
-export class StatusResponseError extends MatterError {
-    public constructor(
-        public readonly message: string,
-        public readonly code: StatusCode,
-    ) {
-        super();
-
-        this.message = `(${code}) ${message}`;
-    }
-}
